+++ Controllers/HomeController.cs
@@ -53,88 +53,6 @@ namespace ForumDemo.Controllers
             return View(vm);
         }
 
-        public async Task<IActionResult> Topic(int id, int? page)
-        {
-            // Pagination
-
-            int pageSize = 20;
-
-            TopicViewModel vm = new TopicViewModel()
-            {
-                Topic = await _topicRepository.GetById(id),
-                Posts = await _topicRepository.GetByIdWithPosts(id, page ?? 1, pageSize)
-            };
-
-            // Manually set breadcrumb nodes
-            var childNode1 = new MvcBreadcrumbNode("Forum", "Home", vm.Topic.Forum.Title)
-            {
-                RouteValues = new { id = vm.Topic.Forum.Id}
-            };
-            var childNode2 = new MvcBreadcrumbNode("Topic", "Home", "ViewData.Title")
-            {
-                OverwriteTitleOnExactMatch = true,
-                Parent = childNode1
-            };
-
-            ViewData["BreadcrumbNode"] = childNode2;
-
-            return View(vm);
-        }
-
-        
-
-        [Authorize]
-        public async Task<IActionResult> Reply(int id)
-        {
-            ReplyViewModel vm = new ReplyViewModel()
-            {
-                Topic =  await _topicRepository.GetById(id)
-            };
-
-            // Manually set breadcrumb nodes
-            var childNode1 = new MvcBreadcrumbNode("Forum", "Home", vm.Topic.Forum.Title)
-            {
-                RouteValues = new { id = vm.Topic.Forum.Id }
-            };
-            var childNode2 = new MvcBreadcrumbNode("Topic", "Home", vm.Topic.Title)
-            {
-                RouteValues = new { id = vm.Topic.Id },
-                OverwriteTitleOnExactMatch = true,
-                Parent = childNode1
-            };
-            var childNode3 = new MvcBreadcrumbNode("Reply", "Home", "Reply")
-            {
-                RouteValues = new { id = vm.Topic.Id },
-                OverwriteTitleOnExactMatch = true,
-                Parent = childNode2
-            };
-
-            ViewData["BreadcrumbNode"] = childNode3;
-
-            return View(vm);
-        }
-
-        [Authorize]
-        [HttpPost]
-        public async Task<IActionResult> Reply(ReplyViewModel vm)
-        {
-            User user = await _userManager.GetUserAsync(HttpContext.User);
-
-            Post post = new Post()
-            {
-                Contents = vm.Contents,
-                Topic = await _topicRepository.GetById(vm.Topic.Id),
-                User = user,
-                Created = DateTime.Now,
-                Updated = DateTime.Now
-            };
-
-            await _postRepository.Create(post);
-            await _userRepository.CountPost(user.Id);
-
-            return RedirectToAction("Topic", new { id = vm.Topic.Id });
-        }
-
         [Authorize]
         public async Task<IActionResult> EditPost(int id)
         {
