+++ Views/Topic/_Pagination.cshtml
@@ -6,7 +6,8 @@
         @if (Model.Posts.HasPreviousPage == true)
         {
             <li class="topic-pagination-button">
-                <a asp-action="Topic"
+                <a asp-controller="Topic"
+                   asp-action="Index"
                    asp-route-id="@Model.Topic.Id"
                    asp-route-page="1">
                     &lt;&lt;
@@ -19,7 +20,8 @@
         @if (Model.Posts.HasPreviousPage == true)
         {
             <li class="topic-pagination-button">
-                <a asp-action="Topic"
+                <a asp-controller="Topic"
+                   asp-action="Index"
                    asp-route-id="@Model.Topic.Id"
                    asp-route-page="@(Model.Posts.PageIndex - 1)">
                     Previous
@@ -34,7 +36,8 @@
             if (Model.Posts.PageIndex - previous_page_button >= 1)
             {
                 <li class="topic-pagination-button">
-                    <a asp-action="Topic"
+                    <a asp-controller="Topic"
+                       asp-action="Index"
                        asp-route-id="@Model.Topic.Id"
                        asp-route-page="@(Model.Posts.PageIndex - previous_page_button)">
                         @(Model.Posts.PageIndex - previous_page_button)
@@ -47,7 +50,8 @@
 
         <li class="topic-pagination-button">
             <a class="current-page"
-               asp-action="Topic"
+               asp-controller="Topic"
+               asp-action="Index"
                asp-route-id="@Model.Topic.Id"
                asp-route-page="@Model.Posts.PageIndex">
                 @Model.Posts.PageIndex
@@ -61,7 +65,8 @@
             if (Model.Posts.PageIndex + next_page_button <= Model.Posts.TotalPages)
             {
                 <li class="topic-pagination-button">
-                    <a asp-action="Topic"
+                    <a asp-controller="Topic"
+                       asp-action="Index"
                        asp-route-id="@Model.Topic.Id"
                        asp-route-page="@(Model.Posts.PageIndex + next_page_button)">
                         @(Model.Posts.PageIndex + next_page_button)
@@ -76,7 +81,8 @@
         @if (Model.Posts.HasNextPage == true)
         {
             <li class="topic-pagination-button">
-                <a asp-action="Topic"
+                <a asp-controller="Topic"
+                   asp-action="Index"
                    asp-route-id="@Model.Topic.Id"
                    asp-route-page="@(Model.Posts.PageIndex + 1)">
                     Next
@@ -89,7 +95,8 @@
         @if (Model.Posts.HasNextPage == true)
         {
             <li class="topic-pagination-button">
-                <a asp-action="Topic"
+                <a asp-controller="Topic"
+                   asp-action="Index"
                    asp-route-id="@Model.Topic.Id"
                    asp-route-page="@Model.Posts.TotalPages">
                     &gt;&gt;
